+++ modules/admin/views/db/index.php
@@ -130,6 +130,18 @@ JS
                     ]
                 ]);
             ?>
+
+            <?
+                echo \mihaildev\elfinder\ElFinder::widget([
+                    'language'         => \Yii::$app->admin->languageCode,
+                    'controller'       => 'cms/elfinder-full', // вставляем название контроллера, по умолчанию равен elfinder
+                    //'filter'           => 'image',    // фильтр файлов, можно задать массив фильтров https://github.com/Studio-42/elFinder/wiki/Client-configuration-options#wiki-onlyMimes
+                    'callbackFunction' => new \yii\web\JsExpression('function(file, id){}'), // id - id виджета
+                    'frameOptions' => [
+                        'style' => 'width: 100%; height: 800px;'
+                    ]
+                ]);
+            ?>
         <? else: ?>
             <p>Дирриктория с файлами бэкапов базы данных не найдена.</p>
         <? endif; ?>